One-quarter-working version! See commit message. We have a parser working for merge-only minimalist grammars. We supposedly implemented both a procedural version and a parallel version, but the procedural one doesn't work. At the same time, the parser only works from within the MinG/S13.pm6 file, for some reason.
It's pretty close to being a success, though the only-inside-the-original-file restriction seems messy. Bumped version to 0.3.0. Version 0.4.0 will be the parallel merge-only parser working from any file (even if I need, say, a wrapper subroutine instead of directly accessing MinG::S13::Parser).

Version 1.0.0 will be when we get a merge-and-move parser working, even if it's just the parallel one.